use essentially shared libfaad

+++ aac/decoder.c
@@ -217,7 +217,7 @@ static int latmCheck(latm_header *latm, bitfile *ld)
     while (ld->bytes_left)
     {
         bits = faad_latm_frame(latm, ld);
-        if (bits==0xFFFFFFFF)
+        if(bits==0xFFFFFFFF)
             bad++;
         else
         {
@@ -1094,7 +1094,7 @@ static void* aac_frame_decode(NeAACDecStruct *hDecoder,
                 faad_free(hDecoder->sample_buffer);
             hDecoder->sample_buffer = NULL;
             hDecoder->sample_buffer = faad_malloc(frame_len*output_channels*stride);
-        } else if (sample_buffer_size < frame_len*output_channels*stride) {
+        } else if (sample_buffer_size < (unsigned long)(frame_len*output_channels*stride)) {
             /* provided sample buffer is not big enough */
             hInfo->error = 27;
             return NULL;
